requirements.txt
Django==1.10.2
djangorestframework==3.5.1
drf-nested-routers==0.11.1
